fix: 切换视图时数据集加载异常
